requirements.txt
altair  >=4.2.2 
attrs  >=22.2.0
blinker  >=1.5
cachetools >=5.3.0 
certifi  >=2022.12.7 
chardet  >=5.1.0 
charset-normalizer >=3.0.1 
click  >=8.1.3 
click-plugins  >=1.1.1 
cligj  >=0.7.2 
dataclass-csv >=1.4.0 
decorator  >=5.1.1 
dnspython  >=2.3.0 
email-validator  >=1.3.1 
entrypoints  >=0.4
fiona  >=1.9.0 
geopandas  >=0.12.2
gitdb  >=4.0.10
gitpython  >=3.1.30
graphviz >=0.20.1
greenlet >=2.0.2 
h3 >=3.7.6 
idna >=3.4
imageio  >=2.25.0
importlib-metadata  >=6.0.0 
jinja2 >= 3.1.2 
jsonschema >=4.17.3
markdown-it-py
markupsafe >=2.1.2 
mdurl
munch  >=2.5.0 
npm  >=0.1.1 
numpy  >=1.24.2
opencv-python  >=4.7.0.68
optional-django  >=0.1.0 
packaging  >=23
pandas >=1.5.3 
passlib  >=1.7.4 
pillow >=9.4.0 
pims >=0.6.1 
protobuf >=3.20.3
psycopg2-binary  >=2.9.5 
pyarrow  >=11.0.0
pydantic >=1.10.4
pydeck >=0.8.0 
pygments >=2.14.0
pympler  >=1.0.1 
pyproj >=3.4.1 
pyrsistent >=0.19.3
python-dateutil  >=2.8.2 
python-dotenv  >=0.21.1
pytz >=2022.7.1
pytz-deprecation-shim  >=0.1.0.post0 
pyyaml >=6
requests >=2.28.2
rich >=13.3.1
semver >=2.13.0
shapely  >=2.0.1 
six  >=1.16.0
slicerator >=1.1.0 
smmap  >=5.0.0 
sqlalchemy >=1.4.46
streamlit  >=1.19.0
toml >=0.10.2
toolz  >=0.12.0
tornado  >=6.2
typing-extensions  >=4.4.0 
tzdata >=2022.7
tzlocal  >=4.2
urllib3  >=1.26.14 
validators >=0.20.0
watchdog >=2.2.1 
zipp >=3.12.1
